Move package to package/manifest
