+++ packages/manifest/__mocks__/npm_dependencies.js
