+++ packages/manifest/config.js
