+++ packages/manifest/dependencies.js
