+++ packages/manifest/index.js
