+++ packages/manifest/npm_dependencies.js
