+++ packages/manifest/packageConfig.js
