+++ packages/manifest/tests/dependencies.test.js
