+++ packages/manifest/tests/npm_dependencies.test.js
