+++ packages/manifest/tests/packageConfig.test.js
