+++ packages/manifest/utils/fetch.js
