+++ packages/manifest/utils/ids.js
