+++ packages/manifest/utils/parse.js
