+++ packages/manifest/utils/parse.test.js
